Created new FAQ section on Landing Page and made new Questions component to make it easier to implement FAQ page and landing page.
Added bottom padding to sign up page to look better
