+++ checkpoint-gamingjournal/src/app/auth/signup/signUp.module.css
@@ -13,6 +13,7 @@
   padding-top: 40px;
   padding-right: 60px;
   padding-left: 40px;
+  padding-bottom: 40px;
   background-color: #303134;
   font-family: 'Poppins';
   justify-content: center;
