+++ checkpoint-gamingjournal/src/app/page.module.css
@@ -172,6 +172,22 @@
   margin-bottom: 6rem;
 }
 
+.faqSection {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 0 20px; /* Padding for the FAQ section */
+  margin-bottom: 3rem;
+}
+
+.faqTitle{
+  font-family: 'Poppins';
+  font-size: 32px;
+  text-align: center;
+  border-bottom: 3px solid white;
+  color: white;
+}
 
 @media (max-width: 1245px) {
   .image{
