+++ checkpoint-gamingjournal/src/components/FrequentQuestions/FrequentQuestions.module.css
@@ -1,0 +1,8 @@
+
+.accordion{
+    display: flex;
+    flex-direction: column;
+    margin-top: 2rem;
+    gap: 20px;
+    font-family: 'Poppins';
+}
